Move the binning objects into their own module
